+++ rules-tests/DowngradePhp70/Rector/New_/DowngradeAnonymousClassRector/config/php_70.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 use Rector\Core\Configuration\Option;
 use Rector\Core\ValueObject\PhpVersionFeature;
-use Rector\DowngradePhp70\Rector\Class_\DowngradeAnonymousClassRector;
+use Rector\DowngradePhp70\Rector\New_\DowngradeAnonymousClassRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
